refactor(curric): tabs into pages and components
